upt_2023.04.07_1

+++ 05.cebir/2. ucgenin_alanini_hesaplama.cpp
@@ -14,7 +14,9 @@ int main()
 
     int yukseklik;
 
-    int alan;
+    //int alan; // Eger sonuc kusuratli bir degerse int en yakin alt degeri alir
+
+    float alan;
 
     cout<< "Ucgenin bir kenarini giriniz: ";
     cin>>kenar;
